Improvement + Reformat Code

+++ src/test/java/ru/netology/stats/StatsServiceTest.java
@@ -11,7 +11,7 @@ public class StatsServiceTest {
         StatsService service = new StatsService();
         long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expectedSum = 180;
-        int actualSum = service.amountPerYear(sales);
+        long actualSum = service.amountPerYear(sales);
 
         Assertions.assertEquals(expectedSum, actualSum);
